refactor: split store.ts into focused submodules (schema, vectors, graph, cache, documents)
Extracts the 3,648-line store.ts god object into 5 single-concern modules under src/store/. The original src/store.ts becomes a thin barrel re-export shim so all existing imports continue to work unchanged.
